메뉴 완료

+++ ddCafe/src/main/java/ddCafe_Manage/MenuUI.java
@@ -2,8 +2,8 @@ package ddCafe_Manage;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
+import java.util.ArrayList;
 import java.util.List;
-
 
 import db.util.DBConn;
 
@@ -11,6 +11,7 @@ import db.util.DBConn;
 public class MenuUI {
 	private BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 	private MenuDAO dao = new MenuDAOmpl();
+	private IngredientDAO dio = new IngredientImpl();
 	
 	public void menu() {
 		
@@ -45,26 +46,99 @@ public class MenuUI {
 
 	public void add_menu() {
 		System.out.println("\n✦ 메뉴 추가 ︎✦");
-		
+		System.out.println("카테고리번호 / 카테고리");
 		try {
+			List<MenuDTO> list = dao.selectCategory();
+			for(MenuDTO dto : list) {
+				System.out.println(dto.getCategoryNum()+". " + dto.getCategoryName());		
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		System.out.println();
+			
+		try {
+			int a,b;
 			MenuDTO dto = new MenuDTO();
 			
 			System.out.print("추가하고싶은 메뉴?");
 			dto.setMenuName(br.readLine());
-			
+		
 			System.out.print("카테고리 번호");
-			dto.setCategoryNum(Integer.parseInt(br.readLine()));
+			b = Integer.parseInt(br.readLine());
+			if(b >7) {
+				System.out.println("카테고리 번호 오류");
+				return;
+			}else {
+				dto.setCategoryNum(b);
+			}			
 			
-			System.out.print("현재 상태?");
-			dto.setStatus(br.readLine());
-			
-			System.out.print("메뉴 가격?");
+			int ch;
+			do {
+				System.out.print("현재 상태 ? [1.주문가능/2.품절]");
+				ch = Integer.parseInt(br.readLine());
+			} while(ch<1||ch>2);
+			String status = null;
+			if(ch==1) {
+				status = "null";
+			} else if(ch==2) {
+				status = "품절";
+			}
+			dto.setStatus(status);
+				
+			System.out.print("메뉴 가격 ? ");
 			dto.setMenuPrice(Integer.parseInt(br.readLine()));
+			int c;
+			do {
+				System.out.print("사이즈 ? [1.R/2.L]");
+				c = Integer.parseInt(br.readLine());
+			} while(c<1||c>2);
+			String size = null;
+			if(c==1) {
+				size = "R";
+			} else if(c ==2) {
+				size = "L";
+			}
+			dto.setMenuSize(size);
+
+			try {
+				List<IngredientDTO> list = dio.left_ingredient();
+				System.out.println("✦ 재료 확인 ︎✦");
+
+				for (IngredientDTO dto1 : list) {
+
+					System.out.println(dto1.getIngredient_code()+ ". "+dto1.getIngredient_name());
+
+				}
+					
+				System.out.println();
+			} catch (NumberFormatException e) {
+				System.out.println("수량은 숫자만 가능합니다");
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+			int d;
+			List<Integer> list = new ArrayList<>();
+			do {
+				System.out.print("메뉴에 맞는 재료번호 ? 0.모든재료 선택완료");
+				d=Integer.parseInt(br.readLine());
+				list.add(d);
+			}while(d!=0);
+			list.remove(list.size()-1);
 			
-			System.out.print("사이즈?");
-			dto.setMenuSize(br.readLine());
+            dto.setIngredients(list);
 			
-			dao.addMenu(dto);
+			
+			System.out.print("정말 추가 하시겠습니까? 1.예 2.아니요");
+			a = Integer.parseInt(br.readLine());
+			if(a==1) {
+				dao.addMenu(dto);
+			}else {
+				return;
+			}
+
+		} catch (NumberFormatException e) {
+			System.out.println("숫자만 가능합니다");
 		} catch (Exception e) {
 			System.out.println("메뉴 등록이 실패했습니다.");
 		}
@@ -72,17 +146,22 @@ public class MenuUI {
 	
 	public void list_menu() {
 		System.out.println("\n메뉴 리스트 !!!");
-		System.out.println("메뉴번호\t메뉴이름\t카테고리번호\t상태\t가격\t사이즈");
+		System.out.println("메뉴번호 / 메뉴이름 / 카테고리번호 / 상태 / 가격 / 사이즈");
 		System.out.println("-----------------------------------------------------------------------");
 		
 		List<MenuDTO> list = dao.listMenu();
 		for(MenuDTO dto : list) {
-			System.out.print(dto.getMenuNum()+"\t");
-			System.out.print(dto.getMenuName()+"\t\t");
-			System.out.print(dto.getCategoryNum()+"\t");
-			System.out.print(dto.getStatus()+"\t");
-			System.out.print(dto.getMenuPrice()+"\t");
+			System.out.print(dto.getMenuNum()+". ");
+			System.out.print(dto.getMenuName()+"/");
+			System.out.print(dto.getCategoryNum()+"/");
+			if(dto.getStatus().equals("품절")) {
+				System.out.print("품절/");
+			} else {
+				System.out.print("주문가능/");
+			}
+			System.out.print(dto.getMenuPrice()+"/");
 			System.out.println(dto.getMenuSize());
+
 		
 		}
 		System.out.println();
@@ -128,19 +207,36 @@ public class MenuUI {
 		for(MenuDTO dto : list) {
 			System.out.println(dto.getSoldMenu());
 		}
+		
 		System.out.println();
 	}
 	
 	public void delete_menu() {
 		System.out.println("\n✦ 메뉴 삭제 ︎✦");
 		
-		int code;
+		int code,c;
+		
+		try {
+			List<MenuDTO> list=dao.showTwoMenu();
+			for(MenuDTO dto : list) {
+				System.out.println(dto.getMenuNum()+". "+dto.getMenuName());
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		System.out.println();
+		
 		try {
 			System.out.print("삭제할 메뉴코드?");
 			code = Integer.parseInt(br.readLine());
-			
-			dao.deleteMenu(code);
 
+			System.out.print("정말 삭제 하시겠습니까? 1.예 2.아니요");
+			c = Integer.parseInt(br.readLine());
+			if(c == 1) {
+				dao.deleteMenu(code);
+			}else {
+				return;
+			}
 			System.out.println("메뉴를 삭제 했습니다.");
 			
 		} catch (Exception e) {
@@ -152,13 +248,27 @@ public class MenuUI {
 	}
 	public void sold_out() {
 		System.out.println("\n 메뉴 품절 처리");
-		int code;
-		
+		int code,c;
+		try {
+			List<MenuDTO> list=dao.showTwoMenu();
+			for(MenuDTO dto : list) {
+				System.out.println(dto.getMenuNum()+". "+dto.getMenuName());
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		System.out.println();
 		try {
 			System.out.println("품절 처리할 메뉴코드?");
 			code = Integer.parseInt(br.readLine());
+			System.out.print("정말 품절처리 하시겠습니까? 1.예 2.아니요");
+			c = Integer.parseInt(br.readLine());
+			if(c == 1) {
+				dao.updateSoldOut(code);
+			}else {
+				return;
+			}
 			
-			dao.updateSoldOut(code);
 			System.out.println("메뉴 품절처리 완료");
 		} catch (Exception e) {
 			System.out.println("품절처리 실패!");
